requirements.txt
transformers==4.14.1
bitsandbytes-cuda111==0.26.0
datasets==1.16.1
loguru
fastapi
uvicorn
torch
flask